[WFLY-7405] messaging-activemq's statistics-enabled is not updated
The write handler to update the statistics-enabled (and
message-counter-sample-period & message-counter-max-day-history)
attribute was using
AttributeDefinition.resolveModelAttribute to get the value but this is
not correct with a :write-attribute operation (where the value is stored
in the "value" node).
Instead, use the newValue from the applyUpdateToRuntime overriden
method.

Update the JMSQueueManagementTestCase#testMessageCounters test to check
that the message counters are actually updated when the server's
statistics-enabled is set to true.

JIRA: https://issues.jboss.org/browse/WFLY-7405